# openapi.yaml
openapi: 3.0.3
info:
  title: Joyna API
  description: >
    API for creating, viewing, updating and sharing events within a user's network.
    Events can be forwarded by invited users to their own networks, and users can
    organize members of their network into groups.
  version: '1.0.0'
servers:
  - url: /
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserPayload'
      responses:
        '200':
          description: User registered and session established
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body or validation error
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: A user already exists with this email
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, session established
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request body
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Invalid credentials
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out and destroy the current session
      operationId: logoutUser
      security: []
      responses:
        '204':
          description: Logged out successfully
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /events:
    get:
      tags: [Events]
      summary: List events for the authenticated user
      operationId: getEvents
      parameters:
        - name: sort
          in: query
          schema:
            $ref: '#/components/schemas/EventSortField'
        - name: order
          in: query
          schema:
            $ref: '#/components/schemas/SortOrder'
        - name: scope
          in: query
          schema:
            $ref: '#/components/schemas/EventListScope'
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Invalid query parameter
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags: [Events]
      summary: Create a new event
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventPayload'
      responses:
        '200':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: >
            Invalid request body, or validation error (empty name, negative spread,
            past event date, owner not found, invalid rsvp deadline, invalid event type)
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /events/{id}:
    patch:
      tags: [Events]
      summary: Update an existing event
      description: Only the event owner may update the event. Omitted fields are left unchanged.
      operationId: updateEvent
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventPayload'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: >
            Invalid request body/id, or validation error (negative spread, past event
            date, invalid rsvp deadline, owner not found)
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not the owner of the event
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Event not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags: [Events]
      summary: Delete an event
      operationId: deleteEvent
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '204':
          description: Event deleted
        '400':
          description: Invalid id
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Event not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /events/invites:
    post:
      tags: [Events]
      summary: Invite (or forward an invite to) a user for an event
      description: >
        If the requester is the event owner, a direct invite is created. Otherwise,
        the requester must already be invited themselves; the invite is then
        forwarded to the new user with no further spread allowed.
      operationId: createEventInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventInvitePayload'
      responses:
        '200':
          description: Invite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventInvite'
        '400':
          description: Invalid request body, invalid eventId/invitedUserId, or negative spread
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User is not allowed to invite (additional) users to this event
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Event not found
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: User has already been invited
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /notifications:
    get:
      tags: [Notifications]
      summary: List the authenticated user's notifications
      description: >
        Returns notifications most-recent-first, with `isRead` reflecting each
        notification's state as of just before this call. As a side effect,
        any currently-unread notifications are then marked read — visiting
        this endpoint is what flips them, so the badge/unread count only
        reflects notifications raised since the last visit.
      operationId: getNotifications
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get the authenticated user's unread notification count
      description: Used to render the badge on the notifications icon. Does not mark anything as read.
      operationId: getUnreadNotificationCount
      responses:
        '200':
          description: Unread notification count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadNotificationCount'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /groups:
    get:
      tags: [Groups]
      summary: List the caller's groups
      description: >
        Returns every group the caller owns, including groups with no members
        yet — unlike deriving groups from GET /network, this surfaces a group
        immediately after creation.
      operationId: listGroups
      responses:
        '200':
          description: The caller's groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags: [Groups]
      summary: Create a new group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupPayload'
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request body, or empty group name
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /groups/{id}:
    patch:
      tags: [Groups]
      summary: Update an existing group
      description: Only the group owner may update the group. Omitted fields are left unchanged.
      operationId: updateGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupPayload'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid request body/id, or empty group name
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Group not found
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: A group with this name already exists
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags: [Groups]
      summary: Delete a group
      operationId: deleteGroup
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '204':
          description: Group deleted
        '400':
          description: Invalid id
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Group not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /network:
    get:
      tags: [Network]
      summary: List the caller's network
      description: >
        Returns every contact the caller has added, each annotated with the group
        it belongs to (contacts without a group belong to the default
        "Acquaintances" bucket in the UI) and how many events the caller and that
        contact have both attended.
      operationId: listConnections
      responses:
        '200':
          description: The caller's connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
    post:
      tags: [Network]
      summary: Add a contact to the caller's network
      operationId: createConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectionPayload'
      responses:
        '200':
          description: Connection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
        '400':
          description: Invalid request body/ids, or attempting to add yourself
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Contact or group not found
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: This contact is already in the caller's network
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /network/potential:
    get:
      tags: [Network]
      summary: List potential connections
      description: >
        Users who attended an event the caller also attended (as owner or
        invitee) and who aren't in the caller's network yet, sorted by number of
        shared events descending.
      operationId: listPotentialConnections
      responses:
        '200':
          description: Potential connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PotentialConnection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /network/lookup:
    get:
      tags: [Network]
      summary: Find a user by exact email match
      description: >
        Backs the "add by email" screen: looks up any registered user (not just
        existing contacts) by an exact email match, so the caller can add them to
        their network. A 404 means no account exists with that email, at which
        point the client offers to invite them to Joyna instead.
      operationId: lookupUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: A user was found with this email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailLookupResult'
        '400':
          description: Missing email query parameter
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No user exists with this email
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
  /network/{contactId}:
    patch:
      tags: [Network]
      summary: Update a connection
      description: >
        Partially updates a connection (its group and/or favorite status). Omitted
        fields are left unchanged; `groupId` set to an empty string clears the
        group back to the default "Acquaintances" bucket.
      operationId: updateConnection
      parameters:
        - $ref: '#/components/parameters/ContactId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConnectionPayload'
      responses:
        '200':
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
        '400':
          description: Invalid request body/id
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Connection or group not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            text/plain:
              schema:
                type: string
components:
  parameters:
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GroupId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ContactId:
      name: contactId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid session
      content:
        text/plain:
          schema:
            type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session-based authentication managed by alexedwards/scs. The cookie is set
        by /auth/register or /auth/login and must be sent on subsequent requests to
        protected routes.
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        joinedAt:
          type: string
          format: date-time
        profilePictureKey:
          type: string
          nullable: true
      required: [id, name, email, joinedAt]
    RegisterUserPayload:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [email, password]
    EventType:
      type: string
      enum: [birthday, party, dinner, gathering, concert, other]
    Mood:
      type: string
      enum: [chill, fun, party, cozy, adventure, romantic, competitive, low-key]
    EventInviteStatus:
      type: string
      enum: [pending, accepted, declined]
    EventSortField:
      type: string
      enum: [date, createdAt]
      default: date
    SortOrder:
      type: string
      enum: [asc, desc]
      default: desc
    EventListScope:
      type: string
      enum: [owned, invited, all]
      default: owned
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        date:
          type: string
          format: date-time
        location:
          type: string
        rsvpDeadline:
          type: string
          format: date-time
          nullable: true
        type:
          $ref: '#/components/schemas/EventType'
        defaultSpreadAllowed:
          type: integer
        mood:
          $ref: '#/components/schemas/Mood'
      required:
        [
          id,
          ownerId,
          name,
          description,
          createdAt,
          date,
          location,
          type,
          defaultSpreadAllowed,
        ]
    CreateEventPayload:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        rsvpDeadline:
          type: string
          format: date-time
          nullable: true
        type:
          $ref: '#/components/schemas/EventType'
        defaultSpreadAllowed:
          type: integer
          minimum: 0
        mood:
          $ref: '#/components/schemas/Mood'
      required: [name, description, date, location, type, defaultSpreadAllowed]
    UpdateEventPayload:
      type: object
      description: All fields are optional; omitted fields are left unchanged.
      properties:
        name:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        rsvpDeadline:
          type: string
          format: date-time
          nullable: true
        type:
          $ref: '#/components/schemas/EventType'
        defaultSpreadAllowed:
          type: integer
          minimum: 0
        mood:
          $ref: '#/components/schemas/Mood'
    EventInvite:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        invitedBy:
          type: string
          format: uuid
        invitedUserId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/EventInviteStatus'
        spreadAllowed:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        [eventId, invitedBy, invitedUserId, status, spreadAllowed, createdAt]
    CreateEventInvitePayload:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        invitedUserId:
          type: string
          format: uuid
        spreadAllowed:
          type: integer
          minimum: 0
      required: [eventId, invitedUserId, spreadAllowed]
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        isFavorite:
          type: boolean
      required: [id, ownerId, name, createdAt, isFavorite]
    CreateGroupPayload:
      type: object
      properties:
        name:
          type: string
      required: [name]
    UpdateGroupPayload:
      type: object
      description: All fields are optional; omitted fields are left unchanged.
      properties:
        name:
          type: string
        isFavorite:
          type: boolean
    NotificationType:
      type: string
      enum:
        [
          event_invite,
          invite_response,
          event_updated,
          rsvp_deadline_reminder,
          event_starting_today,
        ]
    Notification:
      type: object
      description: >
        eventName/actorName are resolved server-side from eventId/actorId at
        read time, so they always reflect the event/user's current name.
        status is only present for invite_response notifications.
        rsvp_deadline_reminder and event_starting_today are raised by a daily
        scheduled job rather than a user action, so they have no actorId.
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/NotificationType'
        eventId:
          type: string
          format: uuid
          nullable: true
        eventName:
          type: string
          nullable: true
        actorId:
          type: string
          format: uuid
          nullable: true
        actorName:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/EventInviteStatus'
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, type, isRead, createdAt]
    UnreadNotificationCount:
      type: object
      properties:
        count:
          type: integer
      required: [count]
    Connection:
      type: object
      properties:
        contactId:
          type: string
          format: uuid
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        isFavorite:
          type: boolean
        groupId:
          type: string
          format: uuid
          nullable: true
        groupName:
          type: string
          nullable: true
        groupIsFavorite:
          type: boolean
          nullable: true
        eventsTogetherCount:
          type: integer
          description: Number of events the caller and this contact have both attended.
      required:
        [
          contactId,
          contactName,
          contactEmail,
          createdAt,
          isFavorite,
          eventsTogetherCount,
        ]
    CreateConnectionPayload:
      type: object
      properties:
        contactId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
      required: [contactId]
    UpdateConnectionPayload:
      type: object
      description: >
        All fields are optional; omitted fields are left unchanged. Setting
        groupId to an empty string clears the group back to "Acquaintances".
      properties:
        groupId:
          type: string
        isFavorite:
          type: boolean
    PotentialConnection:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        sharedEventCount:
          type: integer
      required: [userId, name, email, sharedEventCount]
    EmailLookupResult:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
      required: [userId, name, email]
security:
  - sessionCookie: []